Improved changeShapesWithFeatures
The function solves a total of 24 tasks so far, but can be improved to solve more.

Notice that I got rid of all the "feature"-related things in the Task module. I think that it's better to define the features that are of your interest for each particular application, rather than defining a standard set of features.