Update Task.Shape.hasSameShape
Now this method accepts two new arguments: mirror and scaling.

If mirror is set to True, then the function will return True if the two Shapes are equal modulo mirroring (lr or ud).
If scaling is set to True, then the function will return True if the two Shapes are equal modulo scaling (meaning that they have two different sizes).