Add getPixelChangeCriteria
It serves as substitution of `getPixelFeatures`, which was way too long. The newly added function is easy to understand and fast enough.

Note that I also added a function zoltanRecolor. This function achieves 0.99 in the Kaggle test dataset apparently, so I decided to test it as it is. But I think that `getPixelChangeCriteria` should do the work. I will test it.