Rotate task if necessary
In main, I included the functions `rotateHVTask`, `rotateTaskWithOneBorder` and `recoverRotations`. I also added in the `Task` module the attributes `isVertical` and `isHorizontal` for the classes `Matrix` and `Sample`, as well as the attribute `requiresHVRotation` for the class `Task`.
I also added the attributes `fullBorders` for `Matrix`, `commonFullBorders` for `Sample` and `hasOneFullBorder` for `Task`.