Improve candidate choice criteria
If the task is separated, and there are candidates with score of 0 in both cases (separate and not), I make sure that I choose at least one candidate per mode. I also redefined the order relationship for the class Candidate. If two objects of the class candidate have the same score, then the number of operations breaks the equality.